split readers (llighter imports)
