+++ workflow/scripts/build_renewable_potential.py
@@ -22,7 +22,7 @@ from shutil import rmtree
 import os.path
 
 from _helpers import mock_snakemake, configure_logging
-from readers import read_province_shapes
+from readers_geospatial import read_province_shapes
 from constants import (
     PROV_NAMES,
     CRS,
