+++ workflow/scripts/determine_availability_matrix.py
@@ -29,7 +29,7 @@ import os.path
 from pandas import concat
 from os import mkdir
 
-from readers import read_province_shapes, read_offshore_province_shapes
+from readers_geospatial import read_province_shapes, read_offshore_province_shapes
 from _helpers import configure_logging, mock_snakemake
 from constants import PROV_NAMES, OFFSHORE_WIND_NODES
