+++ workflow/scripts/prepare_base_network.py
@@ -36,7 +36,7 @@ from _pypsa_helpers import (
 )
 from add_electricity import load_costs, sanitize_carriers
 from functions import haversine
-from readers import read_province_shapes
+from readers_geospatial import read_province_shapes
 
 logger = getLogger(__name__)
 logger.setLevel("INFO")
